Rename IDNA backend implementations and refactor loading mechanism
